#1267: Allow to set the sites for which auth strategies will be available
In config/application.yml a domains attribute can be added for each auth
strategy. If not present or with nil value the strategy will be
available for all sites. To restrict the use of the auth strategy for a
group of sites the value of this attribute must be an array containing
their domains.