Add filters to projects API index - First iteration
For the moment the API only accepts nested params in the query string of
the form ?filter[uid]=value&filter[uid2]=value or in a JSON sent in the
body, and the filter creates a query where the custom fields must equal
the values. In the next iteration the filter params will allow intervals
or list of values.