Add meta with statistics to meta action
The fields to generate statistics can be configured at module level
for each site under the key "filterable_custom_fields". The value must
be an array containing the uids of the custom fields. If this
configuration is not present statistics are generated for all custom_fields of
type date, vocabulary_options and numeric.