Allow definition of read only user fields from auth modules
In config/application.yml, at the auth modules section, for an auth
provider a read_only_user_attributes can be defined, with an array of
user attributes identifiers which the user won't be able to edit. The
attributes identifiers can be both names of attributes defined in the
user model or names of custom user fields. The values for these
attributes are expected to be assigned by the session or census custom
forms provided by the auth provider